venv installation
virtualenv has a tendency to not work causes more issues for users. as it will in a lot of cases keep saying the command doesn't exist

so its better to install python3-venv then create a venv with python3 -m venv venv

rest of the instructions still work. but would highly suggest to start recommending pipx instead and to use it with twitch miner as it creates a virtual environment during installation

or just use pipx in general as people will get an error on modern linux systems that say externally-managed-enviorment. which confuses people if you continue reading it python suggests pipx instead now as its techinally safer and cant cause system breakage