Fixes #722
Implements a WebSocket client and pool for the Hermes API.

Abstracts some of the handling for PubSub Messages.

Adds the "use_hermes" Setting to allow the use of the Hermes WebSocket API instead of the PubSub WebSocket API. Defaults to False.

Renamed WebSocket classes and moved them to new websocket module.